Added Azure Warehouse endpoint.
The Warehouse.Azure endpoint publishes the ItemRestocked event and a test Warehouse.Subscriber endpoint to test the handler. The Subscriber has a mutator to remove the assembly information in the enclosed message type so that, the subscriber does not a message dll reference for the scehma and it just declares the same schema as a .cs file in the project.